Basic setup for a lot of classes.
Set up some enums to avoid magic strings, and because certain characters became repetitive. 

Changes in Grid:
 - Horizontal and Vertical were backwards.
 - The ships now have correct size and still display correct output.
 - Complete formatting on grid printing. (Code looks ugly with it all combined though. Can we improve this?)
 - Added more testing units.
 
There exists a 'testing' bug where if same spot was hit twice, it was considered a hit, but this will not occur when the server-sided logic is set up, since it will be tested for validity first. No need to change.