Add files via upload
Added a GameOverException, so when the player sinks the last ship, the grid throws this error to signify the game being over.
Also for demonstration purposes, I added a list of options for us to show to Dr. Barlowe. The option list is:
-show public grid
-show private grid
-quit the game
Also, I made it to where if the player doesnt input something valid it just asks them to do it again. Once all ships are sunk the game ends.
The only way to end the game is through the options menu or by sinking the ships instead of inputting something that were not integers.